[fixed] RequestResourceWeightsBatchCommand
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk@2107 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOCoreORB/src/server/zone/objects/creature/commands/RequestResourceWeightsBatchCommand.h
@@ -69,7 +69,7 @@ public:
 		String value;
 		tokenizer.getStringToken(value);
 
-		uint32 schematicID = Long::valueOf(value);
+		uint32 schematicID = Integer::valueOf(value);
 
 		ManagedReference<CraftingManager* > craftingManager = creature->getZoneServer()->getCraftingManager();
